Исправить заменания после первого ревью
