+++ src/com/yandex/tracker/server/servers/HttpTaskServer.java
@@ -1,8 +1,11 @@
-package com.yandex.tracker.server;
+package com.yandex.tracker.server.servers;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpServer;
+import com.yandex.tracker.server.adapters.DurationAdapter;
+import com.yandex.tracker.server.adapters.LocalDateTimeAdapter;
+import com.yandex.tracker.server.handlers.*;
 import com.yandex.tracker.service.TaskManager;
 
 import java.io.IOException;
@@ -14,6 +17,7 @@ public class HttpTaskServer {
     private final TaskManager taskManager;
     private final Gson gson;
     private HttpServer httpServer;
+    private static final int PORT = 8080;
 
     public HttpTaskServer(TaskManager taskManager) {
         this.taskManager = taskManager;
@@ -24,8 +28,8 @@ public class HttpTaskServer {
                 .create();
     }
 
-    public void start(int port) throws IOException {
-        httpServer = HttpServer.create(new InetSocketAddress(port), 0);
+    public void start() throws IOException {
+        httpServer = HttpServer.create(new InetSocketAddress(PORT), 0);
         httpServer.createContext("/tasks", new TasksHandler(taskManager, gson));
         httpServer.createContext("/subtasks", new SubtasksHandler(taskManager, gson));
         httpServer.createContext("/epics", new EpicsHandler(taskManager, gson));
@@ -33,7 +37,7 @@ public class HttpTaskServer {
         httpServer.createContext("/prioritized", new PrioritizedHandler(taskManager, gson));
 
         httpServer.start();
-        System.out.println("HTTP-сервер запущен на " + port + " порту!");
+        System.out.println("HTTP-сервер запущен на " + PORT + " порту!");
     }
 
     public void stop() {
